+++ models/settings/Consultation.php
@@ -59,7 +59,6 @@ class Consultation implements \JsonSerializable
     public bool $speechRequiresLogin = false;
     public bool $allowMultipleTags = false;
     public bool $allowUsersToSetTags = true;
-    public bool $amendmentsHaveTags = false;
     public bool $openslidesExportEnabled = false;
     public bool $showPrivateNotes = true;